pyproject.toml
[build-system]
requires = ["hatchling>=1.22.0"]
build-backend = "hatchling.build"

[project]
name = "tech-notes-rag-bot"
version = "0.1.0"
description = "Minimal Tech Notes RAG bot: FastAPI RAG service + Telegram bot + OpenRouter."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "chromadb>=0.5",
  "sentence-transformers>=2.6",
  "openai>=1.12",
  "python-dotenv>=1.0",
  "python-telegram-bot>=21.0",
  "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src/app"]


